requirements.txt
requests>=2.32.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
tenacity>=8.2.3
pytest>=8.1.0
openpyxl>=3.1.5
